Add Grouping for ModelsList
Since the groups use different types, there's no good way to use the same SectionedFetchRequest; and in fact if we do it wants to animate the changes between groups which isn't what I want.

So easiest way is just to move the request and ForEach into a sub-view and switch between them based on the Picker.