DecoderTypesList - remove last clever section title extension
The section title extensions are clever, but they've become unnecessary as I've figured out more sensible ways of grouping.

Remove the last one that converted an NSManagedObjectID into a socket title by just having a socket(with:) -> Socket function on the view that uses the viewContext.
The nice thing about this is that when \.socket becomes supported, the function call can be replaced with just section.id!